Enable client credentials auth for CLI use
When Canto needs to be accessed e.g. from the CLI, no Neos user is
active and thus no access token can be used. For those cases, a switch
to client credentials mode is now possible.

Note: With this, the need for appId and appSecret in CantoAssetSource
is already back. :)