Move account/authorization connection step around
The authorization flow redirects from Canto back to Neos. Since we
need to connect the created token with the Neos user account, we
need a valid login.

As soon as the media browser is loaded in an iframe (when opened from
the inspector), this did not work, since browsers don't send cookies
under those circumstances (see link below.)

Instead we redirect to a custom action that in turn redirects "via
HTML", thus the cookie is sent and the connection can be made.

http://concord-consortium.github.io/lara-interactive-api/docs/safari-cookies-redirects/