BUGFIX: Change | to - in asset identifer
The Canto API needs scheme and identifier to identify an asset. Those
were stored in the same string separated by a pipe character. Since
the media browser UI uses the asset (proxy) identifier in the markup,
this resulted in JS errors.

Now a dash is used instead of the pipe, the errors are fixed.

Fixes #6