Un-implement ProvidesOriginalUriInterface
This should speed up interaction with the Canto API a lot, as it
avoids calls to the "direct URI" API endpoint, which creates a signed
CloudFront URL when being called. This may take considerable time.

If needed later, this can be re-implemented in a way that returns a
URI pointing to an action that only calls the Canto API when the URL
is actually being used.