FEATURE: Adds isAuthenticated method to the client
The client always redirects to the authentication notice page when the client is not authenticated. When you are in a different context than a backend module, it is not always the best case to get the redirect. So, this method adds to check if the client is authenticated without the redirect.